+++ src/components/PlantingSites/editor/Details.tsx
@@ -2,12 +2,12 @@ import { Box, Typography, useTheme } from '@mui/material';
 import strings from 'src/strings';
 import { PlantingSite } from 'src/types/Tracking';
 
-export type PlantingSiteDetailsProps = {
+export type DetailsProps = {
   onChange: (id: string, value: unknown) => void;
   site: PlantingSite;
 };
 
-export default function PlantingSiteDetails(props: PlantingSiteDetailsProps): JSX.Element {
+export default function Details(props: DetailsProps): JSX.Element {
   const theme = useTheme();
 
   return (
