+++ src/components/PlantingSites/editor/Editor.tsx
@@ -14,15 +14,15 @@ import useForm from 'src/utils/useForm';
 import Card from 'src/components/common/Card';
 import { toMultiPolygonArray } from 'src/components/Map/utils';
 import PageHeaderWrapper from 'src/components/common/PageHeaderWrapper';
-import PlantingSiteForm, { PlantingSiteStep, PlantingSiteStepType } from './PlantingSiteForm';
-import PlantingSiteDetails from './PlantingSiteDetails';
-import PlantingSiteBoundary from './PlantingSiteBoundary';
-import PlantingSiteExclusions from './PlantingSiteExclusions';
-import PlantingSiteZoneBoundaries from './PlantingSiteZoneBoundaries';
-import PlantingSiteSubzoneBoundaries from './PlantingSiteSubzoneBoundaries';
+import Form, { PlantingSiteStep, PlantingSiteStepType } from './Form';
+import Details from './Details';
+import SiteBoundary from './SiteBoundary';
+import Exclusions from './Exclusions';
+import Zones from './Zones';
+import Subzones from './Subzones';
 import StartOverConfirmation from './StartOverConfirmation';
 
-type PlantingSiteEditorProps = {
+type EditorProps = {
   reloadPlantingSites: () => void;
   site: PlantingSite;
   siteType: SiteType;
@@ -36,7 +36,7 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-export default function PlantingSiteEditor(props: PlantingSiteEditorProps): JSX.Element {
+export default function Editor(props: EditorProps): JSX.Element {
   const { reloadPlantingSites, site, siteType } = props;
   const { activeLocale } = useLocalization();
   const contentRef = useRef(null);
@@ -193,7 +193,7 @@ export default function PlantingSiteEditor(props: PlantingSiteEditorProps): JSX.
           </Typography>
         </Box>
       </PageHeaderWrapper>
-      <PlantingSiteForm
+      <Form
         currentStep={currentStep}
         onCancel={onCancel}
         onSaveAndNext={onSaveAndNext}
@@ -203,21 +203,17 @@ export default function PlantingSiteEditor(props: PlantingSiteEditorProps): JSX.
         className={classes.container}
       >
         <Card style={{ display: 'flex', flexDirection: 'column', flexGrow: 1, marginTop: theme.spacing(4) }}>
-          {currentStep === 'details' && <PlantingSiteDetails onChange={onChange} site={plantingSite} />}
+          {currentStep === 'details' && <Details onChange={onChange} site={plantingSite} />}
           {currentStep === 'site_boundary' && (
-            <PlantingSiteBoundary siteBoundary={siteBoundary} setSiteBoundary={setSiteBoundary} />
+            <SiteBoundary siteBoundary={siteBoundary} setSiteBoundary={setSiteBoundary} />
           )}
           {currentStep === 'exclusion_areas' && (
-            <PlantingSiteExclusions exclusions={exclusions} setExclusions={setExclusions} site={plantingSite} />
+            <Exclusions exclusions={exclusions} setExclusions={setExclusions} site={plantingSite} />
           )}
-          {currentStep === 'zone_boundaries' && (
-            <PlantingSiteZoneBoundaries zones={zones} setZones={setZones} site={plantingSite} />
-          )}
-          {currentStep === 'subzone_boundaries' && (
-            <PlantingSiteSubzoneBoundaries onChange={onChange} site={plantingSite} />
-          )}
+          {currentStep === 'zone_boundaries' && <Zones zones={zones} setZones={setZones} site={plantingSite} />}
+          {currentStep === 'subzone_boundaries' && <Subzones onChange={onChange} site={plantingSite} />}
         </Card>
-      </PlantingSiteForm>
+      </Form>
     </TfMain>
   );
 }
